requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
